zstream: self-tuning queues
This PR removes the `qp_batch_budget` and `qp_queue_length` parameters
from the `zq_params_t` struct that's used to create `zstream_queues.`
Queue lengths and batch budgets still exist, but they're set internally
by the queue implementation. In addition, `ZQ_MAX_BATCH` is raised to
1024 items.

### Motivation

Multithreaded queues need some specific value for queue lengths and
batch sizes, but the impact of these values is largely opaque to
developers without detailed benchmarking.

Queue lengths in particular have been difficult to optimize because they
were formerly the only way to limit the amount of in-flight stream
data. Streams vary in payload density, so queue lengths were set
according to relatively pessimistic assumptions. But as of #18982, the
memory consumed by in-flight payloads is limited independently of queue
operations. Queues now have wider latitude to use throughput-optimized
parameters.

### Benchmark results

I've been investigating what those values should be. Here are the main
findings:

- Longer queues are always better. They facilitate larger batches,
  reduce (proportional) overhead, and smooth irregularities in the
  stream profile and in I/O performance. The magnitude of the
  improvement can be a reduction of up to 15% in total execution time.

- The main cost of longer queues is increased memory consumption, not
  because the storage for queue slots is significant but because long
  queues mean a long pipeline and a correspondingly larger number of
  payloads in memory.

- Despite that, driving the pipeline up against memory limits is helpful
  rather than counterproductive. Ingress throttling makes records
  arrive in a burstier pattern, but long queues also enable the
  pipeline to handle those bursts more smoothly.

- Input throttling happens only when there's a bottleneck downstream of
  the input source. Pipelines are linear, so by definition every stage
  has identical average throughput, and that throughput is set by the
  bottleneck. It's fine for the input stage to spend much of its time
  waiting for memory; in cases where that occurs, higher input
  bandwidth generally wouldn't help.

- The time needed to process a byte of payload data varies by more than
  two orders of magnitude among queues, depending on the work each
  queue is doing. This much variation makes it hard to state rules of
  thumb for how large batches ought to be. Queues need to observe their
  own work rates and set batch sizes accordingly.

- Batch sizes can cause problems both by being too small and by being
  too large. Too small, and many more dispatch loops are needed to farm
  out all the work. Too large, and latency starts to affect overall
  performance.

### Changes in this PR

This PR replaces static batch sizes with batches sized to require 400
microseconds of processing on one thread. To determine the mapping
between time and batch size, each queue accumulates the total cost it
has processed and the total wall-clock time it spent processing, and
calculates the ratio. Both are running totals over the life of the
queue, so the estimate converges early and is largely static.

The 400 microsecond figure is empirical. However, as an absolute
reference point, it should be relatively transportable across systems.

Because batch size is now derived from cost, cost functions carry more
weight than they used to: a cost that doesn't scale with processing
time will mis-size batches. This is documented in `zstream_queue.h`.

All queues are now statically sized at 4096 slots. Longer queues would
be even more performant, but 4096 slots is right about where
diminishing returns start to come into play. Note that this is a fixed
per-queue allocation of the ring buffer and its item storage, made at
queue creation time regardless of load; for `drr_packet_t` items it
works out to about 1.4 MB per queue.

Batch sizes are now capped at 1024 work items, up from 32. In theory,
unlimited batch sizes (that is, batches limited only by the 400us time
quantum) would be optimal. But real-world batch size statistics show a
predominance of modestly-sized batches with a long right tail. Few
batches have the opportunity to become larger than 1024 items.

`ZQ_SLOTS_PER_QUEUE` and `ZQ_MAX_BATCH` remain in `zstream_queue.h`,
but only so that tests and diagnostics can reason about the real
boundaries. They are not caller-tunable.

The queue scoring mechanism used during thread dispatch has been
simplified. A queue's score is now just its number of unclaimed work
items. The more complex scoring used in the past also works fine, but
the extra complexity doesn't yield better performance. I also
investigated backpressure scoring, which incorporates knowledge of the
downstream neighbor's state; that wasn't of benefit either. Stochastic
assignment remains, but it's now calculated with integer arithmetic.

Since queue length is no longer a parameter, the
`STANDARD_INPUT_STACK_Q` and `STANDARD_OUTPUT_STACK_Q` macros are gone,
and `parallel_calc_fletcher4()` no longer takes an argument.

Batch size histograms have been separated out from the
`MONITOR_QUEUES` #define into a separate option,
`SHOW_BATCH_HISTOGRAMS`. `MONITOR_QUEUES` now shows mean and standard
deviation figures for each time interval, sampled on both enqueue and
dequeue. It uses Welford's algorithm, and needs `-lm` for `sqrt()`;
that's why libm has been added to `zstream_LDADD`.

`zstream selftest queue` gains a `queue_batch_tuning` case that runs
queues with deliberately dissimilar cost-to-time relationships,
including ones that pin the derived batch size to each end of its
range. `queue_edge_cases` now sweeps stream lengths that straddle
`ZQ_MAX_BATCH` and `ZQ_SLOTS_PER_QUEUE` instead of the parameters that
no longer exist.

+++ cmd/zstream/zstream_io.c
@@ -100,14 +100,14 @@ static pthread_once_t	dif_init_control = PTHREAD_ONCE_INIT;
 static void
 initialize_memory_tracking(void)
 {
-	ssize_t pagesize = (ssize_t)sysconf(_SC_PAGESIZE);
-	ssize_t pages = (ssize_t)sysconf(_SC_PHYS_PAGES);
+	int64_t pagesize = (int64_t)sysconf(_SC_PAGESIZE);
+	int64_t pages = (int64_t)sysconf(_SC_PHYS_PAGES);
 	if (pagesize < 0 || pages < 0) {
 		warnx("unable to read system memory info");
 		payloads.dif_allowed = UINT64_MAX; /* no limit */
 	} else {
-		uint64_t total_mem = (uint64_t)pagesize * (uint64_t)pages;
-		int64_t flex = total_mem - MEMORY_BASE_CUTOFF;
+		int64_t total_mem = pagesize * pages;
+		int64_t flex = total_mem - (int64_t)MEMORY_BASE_CUTOFF;
 		int64_t addl = (double)flex * MEMORY_PCT / 100;
 		payloads.dif_allowed = MEMORY_BASE + MAX(addl, 0);
 	}
